refactoring auth reducers
