--- client/src/reducers/authorization/index.js
@@ -1,4 +1,0 @@
-import authorizationReducer from "./reducer";
-export { default as authorizationTypes } from "./types";
-export { default as authorizationActions } from "./actions";
-export default authorizationReducer;
