--- client/src/reducers/authorization/types.js
@@ -1,13 +1,0 @@
-const LOGIN = "LOGIN";
-const LOGOUT = "LOGOUT";
-const VERIFICATION = "VERIFICATION";
-const REFRESH_TOKEN = "REFRESH_TOKEN";
-
-const ex = {
-  LOGIN,
-  LOGOUT,
-  VERIFICATION,
-  REFRESH_TOKEN,
-};
-
-export default ex;
