+++ client/src/reducers/tokens/index.js
@@ -1,4 +1,4 @@
 import tokensReducer from "./reducer";
-export { default as tokensTypes } from "../../actions/tokens/types";
-export { default as tokensActions } from "../../actions/tokens/actions";
+export { default as tokensTypes } from "./types";
+export { default as tokensActions } from "./actions";
 export default tokensReducer;
